+++ java-vtl-engine/src/main/java/fr/insee/vtl/engine/exceptions/InvalidTypeException.java
@@ -1,4 +1,4 @@
-package fr.insee.vtl.engine;
+package fr.insee.vtl.engine.exceptions;
 
 import org.antlr.v4.runtime.tree.ParseTree;
